Move keybindingTypes to schemas/
