+++ src/constants/coreSettings.ts
@@ -2,7 +2,7 @@ import { LinkMarkerShape } from '@comfyorg/litegraph'
 import { LiteGraph } from '@comfyorg/litegraph'
 
 import type { ColorPalettes } from '@/schemas/colorPaletteTypes'
-import type { Keybinding } from '@/types/keyBindingTypes'
+import type { Keybinding } from '@/schemas/keyBindingTypes'
 import { NodeBadgeMode } from '@/types/nodeSource'
 import { LinkReleaseTriggerAction } from '@/types/searchBoxTypes'
 import type { SettingParams } from '@/types/settingTypes'
