+++ src/schemas/keyBindingTypes.ts
